Query OpenWeatherAPI for information
Application now queries the API for the data. I've added a third state in the weather machine `query`.

the new query state is the next step after the `start` state and is responsible for invoking the service to get the weather information. After that is successfully done, it transitions to the "final" state of display.

Installed Axios in order to perform the API Query, and i've added the OpenWeatherAPI API key to a .env file that gets included by Razzle on Build