Query API by single city name
Implemented a new states in the machine. now from the initial state, an interpreter must determine if there are cities in the query string or not and transition the machine to the appropriate state.

The machine service that gets the API data handles determines dynamically using context if it should call data with geolocation or city name.

currently only works assuming one city. will need to make it work with multiple cities and rotate the requests