Fix periodic weather updates
Fixed an issue that caused the periodic updates to not work properly.

What happened was both initial requests by geolocation and by city were working properly. in case the user was getting data by city, initial transition was caused by the `QUERY` event. however all future periodic updates were handled by the `TICK` event.

This was a problem because the city information was being sent to the request service through event data and not through context. The solution was to refactor this bit in order to make sure the city information was being assigned properly to the context before calling the request service