Change Weathermachine context structure to fit multiple cities
changed the context interface of the weather machine to fit handling multiple city weather data at the same time.

essentially it switched from being a list of cities, plus weather data for one city, to be a list of city objects that includes their own weather data. that was needed in order to hold/cache weather data for multiple locations at the same time.

This needed sweeping changes throughout the display component, the weather machine, and the unit tests to make sure that everything works correctly as it had before