Created stories for Display Component and Start Component
it should be noted that the display component has two stories, one for a single city and the other for multiple cities. The multiple cities story has a control for selecting which city is currently displayed.

i attempted to make it cycle every 5 seconds like the actual controller, however i ran into issues trying to properly bind to the arguments passed to the component in the story. in order to make it programatically update like that was going to be very ugly, so i opted to simply make a custom manual control for the user to pick the current city