Fix mouse selection of images displaying orange color
The images being used are SVG files (ever since i got rid of the font family). One of the limitations of using an SVG file is that you cannot color it through CSS (it has to be inline SVG in order to do that)

So since the SVG files are originally black, i decided to pass them through a CSS filter (color inversion) to get them to white. i could have simply colored the source files, but i still do need a black variant of them (for the background pattens).

The CSS filter worked great and the icons are white. the only issue was that if you selected them by mouse (as you would text), the default blue color would be inverted to orange while selecting the image. this commit basically removes the selection color on all images