Skip initial geolocation state if query contains cities
When loading the page with query string information, the start component would flash before being replaced with the display component. even though the state machine was handling the transition correctly, the issue was with the logic that displays the components (in the `Home` component).

Now the Start component will only be displayed if the state machine is explicitly asking for geolocation