Partial work on video backgrounds
Created a new machine to handle backgrounds of each city. It was close to being finished when i discovered that the youtube API has a strict limit for requests.

Even if the limit gets refreshed every 24 hours, i don't think it's a good idea to depend on it for this project as one could easily enter 5 different cities, each triggering a search request on the API. multiply that by how many times the page would get refreshed and it's relatively easy to hit the limit.

The suggested approach online is to use a different service (Noembed) to search youtube without using an API key and sidestep these kinds of issues.

I'll need to change how i parsed the information from the official API. With the way forward clear, i decided to commit the partial work done and focus my attention elsewhere on something that may yield results faster considering the deadline