Fix noticable jump in scrolling icon background
The original issue was that the scrolling icon background recenters every 5 seconds so as not to expose plain background. this is un-noticable while cycling through cities but obvious with only one city.  Issue was caused by unequal momentum between the x and y axis while scrolling.

The simplest fix for this was to make sure that the icons scrolled in predictable directions and not have full 360 control. So instead of calculating the exact X and Y momentum based on the wind direction, the calculation method now rounds the direction to the nearest 45 degree.

This ensures that by the end of the animation, the scrolling UI icons will always be overlapping properly once the animation restarts and there won't be any percievable "Jump" in the animation